Splits SBPF program test crate `mem` into `dep`.
