+++ programs/sbf/rust/mem_dep/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "solana-sbf-rust-membuiltins"
+name = "solana-sbf-rust-mem-dep"
 version = { workspace = true }
 description = { workspace = true }
 authors = { workspace = true }
@@ -10,7 +10,6 @@ edition = { workspace = true }
 
 [dependencies]
 solana-program = { workspace = true }
-solana-sbf-rust-mem-dep = { workspace = true }
 
 [lib]
-crate-type = ["cdylib"]
+crate-type = ["lib"]
